feat: add to threshold entity if it exists when creating or updating an expense

+++ services/expenses-service/src/main/java/ir/snp/threashold/entity/ExpenseThreshold.java
@@ -6,6 +6,8 @@ import jakarta.persistence.*;
 import jakarta.validation.Valid;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 public class ExpenseThreshold {
@@ -20,8 +22,8 @@ public class ExpenseThreshold {
     @ManyToOne
     private Category category;
 
-    private double monthlyThreshold;
-    private double totalMonthlyExpenses;
+    private BigDecimal monthlyThreshold;
+    private BigDecimal totalMonthlyExpenses;
 
 
 }
